Rename html/svg autogenerated files
