+++ README.md
@@ -27,9 +27,9 @@ browser.close
 Description
 -----------
 
-The file in `lib/watir-webdriver/elements/html.rb` is autogenerated from the HTML spec. This is done by extracting the IDL parts from the spec and processing them with the WebIDL gem (link below).
+The file in `lib/watir-webdriver/elements/html_generated.rb` is autogenerated from the HTML spec. This is done by extracting the IDL parts from the spec and processing them with the WebIDL gem (link below).
 
-The same applies to `lib/watir-webdriver/elements/svg.rb`.
+The same is true `lib/watir-webdriver/elements/svg_elements.rb`.
 
 Specs
 -----
