+++ lib/watir-webdriver.rb
@@ -87,8 +87,8 @@ require 'watir-webdriver/user_editable'
 require 'watir-webdriver/element_collection'
 require 'watir-webdriver/elements/element'
 
-require 'watir-webdriver/elements/html'
-require 'watir-webdriver/elements/svg'
+require 'watir-webdriver/elements/html_elements'
+require 'watir-webdriver/elements/svg_elements'
 
 require 'watir-webdriver/elements/area'
 require 'watir-webdriver/elements/button'
