+++ lib/watir-webdriver/elements/svg_elements.rb
